+++ src/Products/PloneMeeting/MeetingItem.py
@@ -3302,13 +3302,11 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            signatory uid as key and 'signature_number' as value.
            Else, the key is the signatory contact object.
         '''
-        res = []
+        signatories = {}
         if not self.hasMeeting():
-            return res
+            return signatories
         meeting = self.getMeeting()
-        if real:
-            signatories = {}
-        else:
+        if not real:
             signatories = meeting.getSignatories(by_signature_number=True)
         item_signatories = meeting.getItemSignatories().get(self.UID(), {})
         signatories.update(item_signatories)
@@ -4453,7 +4451,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
         data = {}
         tool = api.portal.get_tool('portal_plonemeeting')
-        adviser_org_uids = [org.UID() for org in tool.get_orgs_for_user(suffixes=['advisers'])]
+        adviser_org_uids = tool.get_orgs_for_user(suffixes=['advisers'], the_objects=False)
         for adviceInfo in self.adviceIndex.values():
             advId = adviceInfo['id']
             # if advice is inherited get real adviceInfo